Remove license TODO from readme
GitHub really wants you to put your license in the root directory of the repo. I initially set it to be placed in the documents directory but GitHub automatically changed it to the root directory and I committed it before I noticed, so I am just going to leave it there for now.